Check that directory exists before deleting

+++ vivo-x300-3-extract-live-photos.sh
@@ -64,8 +64,12 @@ while read -r file; do
 	fi
 done <<< "$(find "$INPUT_DIR" -type f -name '*.jpg' -printf '%P\n')"
 
-find "$OUTPUT_VIDEO_DIR" -type d -empty -delete
-find "$OUTPUT_PHOTO_DIR" -type d -empty -delete
+if [ -d "$OUTPUT_VIDEO_DIR" ]; then
+	find "$OUTPUT_VIDEO_DIR" -type d -empty -delete
+fi
+if [ -d "$OUTPUT_PHOTO_DIR" ]; then
+	find "$OUTPUT_PHOTO_DIR" -type d -empty -delete
+fi
 
 echo ""
 if [ ${#motion_photos[@]} -ne 0 ]; then
